Non-sprint task - minor refactoring: introduce generics for component system + avoid unnecessary casts.

+++ org.eclipse.jubula.rc.javafx/src/org/eclipse/jubula/rc/javafx/listener/KeyAcceptor.java
@@ -123,7 +123,7 @@ public class KeyAcceptor {
     public static int getAWTKeyCode(KeyEvent event) {
         KeyCode code = event.getCode();
         if (code.isLetterKey()) {
-            return java.awt.event.KeyEvent.getExtendedKeyCodeForChar((int) code
+            return java.awt.event.KeyEvent.getExtendedKeyCodeForChar(code
                     .getName().charAt(0));
         } else if (code.isFunctionKey()) {
             switch (code) {
